Documentation update for SimpleHTTPResolver
Update the documentation for the new resolver. In addition, fix any
errors I am aware of in it. And add functionality to move the Simple
HTTP resolver clean script to where the other clean script is place.

These are the instructions I had to use for it to run under Ubuntu
14.04.