Accept config options "cert" and "key". (#245)
* Accept config options "cert" and "key".

These values are used to set the cert tuple for requests. In the SimpleHTTPResolver, you may have one cert and one key configured. In the TemplateHTTPResolver, you may have a default cert and key configured in the [resolver] config section, and you may override this default on a per-template basis by putting a cert and key in the appropriate [[template-name]] subsection.

* Updated docstrings in resolver.py.

Added SSL cert authentication info, the new structure of the TemplateHTTPResolver config, and the ssl_check option.