Set the Content-Disposition header in image responses (#476)
Currently browsers will use the last part of the path as a filename,
which is almost always something unhelpful like "default.jpg" or
"default.png".  This patch changes Loris to start sending a header
which tells browsers to include the ident in the responses.

This should make it easier for people to trace images back to the
original from a downloaded file.