Test that bad JWT tokens are rejected by RulesAuthorizer
The `except AuthorizerException as e` branch is now tested, and exposed
a bug when throwing the AuthorizerException -- it doesn't take keyword
arguments.

Additionally, `is_authorized()` will now catch any error thrown by
`_roles_from_request()` and treat it as a denied request -- e.g. if the
origin on the token and the request don't match.