Remove named pipe (fifo) requirements
Using a named pipe for the transforms might be faster. However, the
named pipe requires the user knowing what command to use to create it
(if the default doesn't work). The named pipe also results in the
transform just hanging if there's an error, instead of returning the
error right away.